refactor: move order related hooks to its own entity
