+++ packages/api/src/hooks/index.ts
@@ -22,7 +22,7 @@ export { default as useCashierFiatAddress } from './useCashierFiatAddress';
 // TODO: Export these into P2P entities
 export { default as useChatCreate } from './p2p/useChatCreate';
 export { default as useAdvertiserInfo } from './p2p/useAdvertiserInfo';
-export { default as useOrderInfo } from './p2p/useOrderInfo';
+export { default as useOrderInfo } from './p2p/entity/order/p2p-order/useOrderInfo';
 
 export { default as useCountryConfig } from './useCountryConfig';
 export { default as useCreateNewRealAccount } from './useCreateNewRealAccount';
