+++ packages/api/src/hooks/p2p/entity/order/p2p-order/useOrderInfo.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
-import useQuery from '../../useQuery';
-import useAuthorize from '../useAuthorize';
+import useQuery from '../../../../../useQuery';
+import useAuthorize from '../../../../useAuthorize';
 
 /** This custom hook that returns information about the given order ID */
 const useOrderInfo = (id: string) => {
